+++ confik-macros/src/lib.rs
@@ -8,8 +8,8 @@ use darling::{
 use proc_macro2::{Ident, TokenStream};
 use quote::{format_ident, quote, quote_spanned};
 use syn::{
-    parse_macro_input, spanned::Spanned, DeriveInput, Expr, Generics, Index, Lit, Meta, Path, Type,
-    Visibility,
+    parse2, parse_macro_input, spanned::Spanned, DeriveInput, Expr, Generics, Index, Meta, Path,
+    Type, Visibility,
 };
 
 #[cfg(test)]
@@ -33,17 +33,16 @@ struct FieldFrom {
 }
 
 impl FromMeta for FieldFrom {
-    fn from_value(ty: &Lit) -> darling::Result<Self> {
-        let Lit::Str(ty) = ty else {
-            return Err(darling::Error::unexpected_lit_type(ty).with_span(&ty.span()));
-        };
-
-        let Ok(ty) = ty.parse::<Type>() else {
+    fn from_expr(ty: &Expr) -> darling::Result<Self> {
+        let Ok(ty) = parse2(ty.to_token_stream()) else {
             return Err(syn::Error::new(
                 ty.span(),
-                format!("Unable to parse provided from as rust type: {}", ty.value()),
-            ))
-            .map_err(Into::into);
+                format!(
+                    "Unable to parse provided from ({}) as rust type",
+                    ty.to_token_stream()
+                ),
+            )
+            .into());
         };
 
         Ok(Self { ty })
@@ -82,7 +81,7 @@ impl FromMeta for ForwardSerde {
 /// Parser for a default attribute
 #[derive(Debug)]
 struct FieldDefaulter {
-    expr: TokenStream,
+    expr: Expr,
 }
 
 impl FromMeta for FieldDefaulter {
@@ -92,23 +91,9 @@ impl FromMeta for FieldDefaulter {
         })
     }
 
-    fn from_value(default: &Lit) -> darling::Result<Self> {
-        let Lit::Str(default) = default else {
-            return Err(darling::Error::unexpected_lit_type(default).with_span(&default.span()));
-        };
-
-        let Ok(default) = default.parse::<Expr>() else {
-            return Err(syn::Error::new(
-                default.span(),
-                format!(
-                    "Unable to parse provided default as rust expression: {}",
-                    default.value()
-                ),
-            ))
-            .map_err(Into::into);
-        };
-
-        let expr = quote_spanned!(default.span() => { #default }.into() );
+    fn from_expr(default: &Expr) -> darling::Result<Self> {
+        let default = quote_spanned!(default.span() => { #default }.into() );
+        let expr = parse2(default).expect("Failed to parse default when wrapped in into");
         Ok(Self { expr })
     }
 }