+++ confik-macros/tests/trybuild/fail-default-empty.rs
@@ -1,7 +1,0 @@
-#[derive(confik::Configuration)]
-struct Config {
-    #[confik(default = "")]
-    _param: String,
-}
-
-fn main() {}