+++ confik-macros/tests/trybuild/fail-default-parse.rs
@@ -1,6 +1,6 @@
 #[derive(confik::Configuration)]
 struct Config {
-    #[confik(default = "1 + 2")]
+    #[confik(default = 1 + 2)]
     _param: usize,
 }